+++ src/main/java/com/spotinst/sdkjava/model/ApiGroupCompute.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.spotinst.sdkjava.client.rest.IPartialUpdateEntity;
+import com.spotinst.sdkjava.model.bl.elastigroup.aws.ElastigroupVolumeAttachments;
 
 import java.util.HashSet;
 import java.util.List;
