+++ src/main/java/com/spotinst/sdkjava/model/api/elastigroup/aws/ApiElastigroupVolumeAttachments.java
@@ -6,7 +6,6 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.spotinst.sdkjava.client.rest.IPartialUpdateEntity;
-import com.spotinst.sdkjava.model.ElastigroupVolume;
 
 import java.util.HashSet;
 import java.util.List;
