+++ src/main/java/com/spotinst/sdkjava/model/bl/elastigroup/aws/ElastigroupVolumeAttachments.java
@@ -1,4 +1,4 @@
-package com.spotinst.sdkjava.model;
+package com.spotinst.sdkjava.model.bl.elastigroup.aws;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
